# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CategoryCreatable:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  models.response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      statusCode:
        type: integer
    type: object
host: localhost:3001
info:
  contact:
    email: dso.intern.xuanhoa@gmail.com
    name: Xuan Hoa Le
  description: Go Food Delivery CMS - Swagger API Discovery
  title: Go Food Delivery CMS - Swagger API Discovery
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Read Category
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read Category
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create Category
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryCreatable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Create Category
      tags:
      - categories
  /employees:
    get:
      consumes:
      - application/json
      description: Read All Employee
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read All Employee
      tags:
      - employees
  /employees/{employeesId}:
    delete:
      consumes:
      - application/json
      description: Delete Employee
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Employee Id
        in: path
        name: employeeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Delete Employee
      tags:
      - employees
  /products:
    get:
      consumes:
      - application/json
      description: Read Product By Category Id
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Table Id
        in: path
        name: tableId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read Product By Category Id
      tags:
      - products
  /products/{productId}:
    delete:
      consumes:
      - application/json
      description: Delete Product By Id
      parameters:
      - description: Product Id
        in: path
        name: productId
        required: true
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Delete Product By Id
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Read Product By Id
      parameters:
      - description: Product Id
        in: path
        name: productId
        required: true
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read Product By Id
      tags:
      - products
  /products/category/{categoryId}:
    get:
      consumes:
      - application/json
      description: Read Product By Category Id
      parameters:
      - description: Category Id
        in: path
        name: categoryId
        required: true
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read Product By Category Id
      tags:
      - products
  /products/recommend/{limit}:
    get:
      consumes:
      - application/json
      description: Read Recommend Product
      parameters:
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.response'
      summary: Read Recommend Product
      tags:
      - products
...
swagger: "2.0"